press h to despawn everything
