+++ clientnew/src/main.rs
@@ -1,11 +1,13 @@
 pub mod cameras;
+pub mod states;
+pub mod menu;
 
 use bevy::{
     diagnostic::FrameTimeDiagnosticsPlugin,
     prelude::*,
     window::{Cursor, CursorGrabMode},
 };
-use cameras::spawn_player_sprite;
+use states::GameState;
 
 pub const HEIGHT: f32 = 720.0;
 pub const WIDTH: f32 = 1280.0;
@@ -44,73 +46,18 @@ fn main() {
                 .set(ImagePlugin::default_nearest()),
             cameras::CameraPlugin,
             shared::ConfigPlugin,
+            states::StatePlugin,
+            menu::MenuPlugin,
         ))
-        .add_systems(Startup, (spawn_scene, spawn_player_sprite))
+        .add_systems(Update, bevy::window::close_on_esc) // Close the window when you press escape
         .run();
 }
 
-pub fn spawn_scene(
+pub fn despawn_all_component<T: Component>(
+    items: Query<Entity, With<T>>,
     mut commands: Commands,
-    mut meshes: ResMut<Assets<Mesh>>,
-    mut materials: ResMut<Assets<StandardMaterial>>,
-    asset_server: ResMut<AssetServer>,
 ) {
-    let size = 30.;
-    // Ground
-    commands
-        .spawn((
-            PbrBundle {
-                mesh: meshes.add(Mesh::from(shape::Plane {
-                    size: size * 2.0,
-                    subdivisions: 10,
-                })),
-                material: materials.add(Color::hex("#1f7840").unwrap().into()),
-                transform: Transform::from_xyz(0.0, -0.01, 0.0),
-                ..default()
-            },
-            Name::new("Plane"),
-        ));
-        //.with_children(|commands| {
-            //commands.spawn((
-                //Collider::cuboid(size, 1., size),
-                //Name::new("PlaneCollider"),
-                //TransformBundle::from(Transform::from_xyz(0., -1., 0.)),
-            //));
-        //});
-    // Sun
-    commands.spawn((
-        DirectionalLightBundle {
-            transform: Transform::from_rotation(Quat::from_rotation_x(
-                -std::f32::consts::FRAC_PI_2,
-            ))
-            .mul_transform(Transform::from_rotation(Quat::from_rotation_y(
-                -std::f32::consts::FRAC_PI_4,
-            ))),
-            directional_light: DirectionalLight {
-                color: Color::rgb(1.0, 0.9, 0.8),
-                illuminance: 15_000.0,
-                shadows_enabled: true,
-                ..default()
-            },
-            ..Default::default()
-        },
-        Name::new("Sun"),
-    ));
-    // House
-    commands
-        .spawn((
-            SceneBundle {
-                scene: asset_server.load("sprytilebrickhouse.gltf#Scene0"),
-                transform: Transform::from_xyz(-5.2, -1.0, -20.0)
-                    .with_rotation(Quat::from_rotation_y(std::f32::consts::PI)),
-                ..default()
-            },
-            Name::new("House"),
-        ))
-        .with_children(|commands| {
-            commands.spawn((
-                SpatialBundle::from_transform(Transform::from_xyz(-5., 0., -5.)),
-                //Collider::cuboid(5., 1.0, 6.),
-            ));
-        });
+    for item in &items {
+        commands.entity(item).despawn_recursive();
+    }
 }
